requirements.txt
beautifulsoup4==4.13.5
img2table==1.4.2
llvmlite==0.45.0
numba==0.62.0
numpy==2.2.6
opencv-contrib-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
pyarrow==21.0.0
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-telegram-bot==21.8
pytz==2025.2
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
xlsxwriter==3.2.9
